+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 1.x.x - 2019-##-
 
-1.
+1. Include all dash-looking characters when sanitizing ids in `id-util.js`
 
 ## 0.10.0 - 2019-04-01